server/pyproject.toml
[project]
name = "data-platform-server"
version = "0.1.0"
description = "FastAPI server + Celery workers for the data platform"
requires-python = ">=3.13"
readme = "AGENTS.md"

dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "asgi-correlation-id>=4.3.4",
    "slowapi>=0.1.9",

    # Validation / settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "email-validator>=2.2.0",

    # DB
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.3",            # used by alembic + celery result backend (sync)

    # Cache / queue
    "redis>=5.2.0",
    "celery[redis]>=5.4.0",
    "sqlalchemy-utils>=0.41.2",          # for celery db result backend helpers

    # Vector store / LLM
    "qdrant-client>=1.12.0",
    "openai>=1.55.0",
    "tiktoken>=0.8.0",

    # Logging
    "structlog>=24.4.0",

    # Observability
    "sentry-sdk[fastapi,celery,sqlalchemy]>=2.19.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-celery>=0.50b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.50b0",
    "opentelemetry-instrumentation-httpx>=0.50b0",
    "opentelemetry-instrumentation-redis>=0.50b0",
    "opentelemetry-instrumentation-logging>=0.50b0",
    "prometheus-fastapi-instrumentator>=7.0.0",

    # HTTP client (for outbound calls — keeps OTEL httpx instrumentation effective)
    "httpx>=0.28.0",

    # Config file parsing — YAML is the project's chosen config format.
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "types-redis>=4.6.0",
    "types-pyyaml>=6.0.12",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ── Ruff ──────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = [
    "E", "F", "W",      # pycodestyle + pyflakes
    "I",                # isort
    "B",                # bugbear
    "C4",               # comprehensions
    "UP",               # pyupgrade
    "SIM",              # simplify
    "RUF",              # ruff-specific
    "ASYNC",            # async lints — catches blocking calls in async context
    "TID",              # tidy imports
    "PIE",
    "PL",               # pylint subset
    "TRY",              # tryceratops — catches bad except patterns
    "ARG",              # unused-args
    "PTH",              # use pathlib
    "ANN",              # type annotation completeness
]
ignore = [
    "E501",             # line length handled by formatter
    "PLR0913",          # too many arguments — fastapi deps make this noisy
    "TRY003",           # long messages in raise — fine
    "ANN401",           # `Any` in type annotation — we ban Any via mypy instead
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN", "PLR2004"]
"alembic/env.py" = ["ANN"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ── Mypy ──────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.13"
strict = true
warn_unreachable = true
warn_no_return = true
disallow_any_explicit = true             # bans `Any` in annotations — enforces the "no Any" rule
disallow_any_generics = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]
exclude = ["alembic/versions/", "tests/"]

[[tool.mypy.overrides]]
module = ["celery.*", "slowapi.*", "qdrant_client.*"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# ── Pytest ────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
